Rename to ohs-player-reference-app
