--- client-app/src/androidMain/kotlin/dev/ohs/player/reference/client/app/Platform.android.kt
@@ -1,9 +1,0 @@
-package dev.ohs.player.reference.client.app
-
-import android.os.Build
-
-class AndroidPlatform : Platform {
-    override val name: String = "Android ${Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()
