--- client-app/src/commonMain/kotlin/dev/ohs/player/reference/client/app/Greeting.kt
@@ -1,9 +1,0 @@
-package dev.ohs.player.reference.client.app
-
-class Greeting {
-    private val platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
