--- client-app/src/commonMain/kotlin/dev/ohs/player/reference/client/app/Platform.kt
@@ -1,7 +1,0 @@
-package dev.ohs.player.reference.client.app
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
