--- client-app/src/commonTest/kotlin/dev/ohs/player/reference/client/app/ComposeAppCommonTest.kt
@@ -1,12 +1,0 @@
-package dev.ohs.player.reference.client.app
-
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class ComposeAppCommonTest {
-
-    @Test
-    fun example() {
-        assertEquals(3, 1 + 2)
-    }
-}
