--- client-app/src/iosMain/kotlin/dev/ohs/player/reference/client/app/MainViewController.kt
@@ -1,5 +1,0 @@
-package dev.ohs.player.reference.client.app
-
-import androidx.compose.ui.window.ComposeUIViewController
-
-fun MainViewController() = ComposeUIViewController { App() }
