--- client-app/src/iosMain/kotlin/dev/ohs/player/reference/client/app/Platform.ios.kt
@@ -1,9 +1,0 @@
-package dev.ohs.player.reference.client.app
-
-import platform.UIKit.UIDevice
-
-class IOSPlatform: Platform {
-    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
-
-actual fun getPlatform(): Platform = IOSPlatform()
