--- client-app/src/jsMain/kotlin/dev/ohs/player/reference/client/app/Platform.js.kt
@@ -1,7 +1,0 @@
-package dev.ohs.player.reference.client.app
-
-class JsPlatform: Platform {
-    override val name: String = "Web with Kotlin/JS"
-}
-
-actual fun getPlatform(): Platform = JsPlatform()
