--- client-app/src/jvmMain/kotlin/dev/ohs/player/reference/client/app/Platform.jvm.kt
@@ -1,7 +1,0 @@
-package dev.ohs.player.reference.client.app
-
-class JVMPlatform: Platform {
-    override val name: String = "Java ${System.getProperty("java.version")}"
-}
-
-actual fun getPlatform(): Platform = JVMPlatform()
