--- client-app/src/wasmJsMain/kotlin/dev/ohs/player/reference/client/app/Platform.wasmJs.kt
@@ -1,7 +1,0 @@
-package dev.ohs.player.reference.client.app
-
-class WasmPlatform: Platform {
-    override val name: String = "Web with Kotlin/Wasm"
-}
-
-actual fun getPlatform(): Platform = WasmPlatform()
