--- client-app/src/webMain/kotlin/dev/ohs/player/reference/client/app/main.kt
@@ -1,11 +1,0 @@
-package dev.ohs.player.reference.client.app
-
-import androidx.compose.ui.ExperimentalComposeUiApi
-import androidx.compose.ui.window.ComposeViewport
-
-@OptIn(ExperimentalComposeUiApi::class)
-fun main() {
-    ComposeViewport {
-        App()
-    }
-}
