+++ iosApp/Configuration/Config.xcconfig
@@ -1,7 +1,7 @@
 TEAM_ID=
 
 PRODUCT_NAME=Ohsplayerreferenceclientapp
-PRODUCT_BUNDLE_IDENTIFIER=dev.ohs.player.reference.client.app.Ohsplayerreferenceclientapp$(TEAM_ID)
+PRODUCT_BUNDLE_IDENTIFIER=dev.ohs.player.reference.app.Ohsplayerreferenceclientapp$(TEAM_ID)
 
 CURRENT_PROJECT_VERSION=1
 MARKETING_VERSION=1.0
