+++ ohs-player-reference-app/src/androidMain/AndroidManifest.xml
