+++ ohs-player-reference-app/src/androidMain/kotlin/dev/ohs/player/reference/app/Platform.android.kt
@@ -1,0 +1,9 @@
+package dev.ohs.player.reference.app
+
+import android.os.Build
+
+class AndroidPlatform : Platform {
+    override val name: String = "Android ${Build.VERSION.SDK_INT}"
+}
+
+actual fun getPlatform(): Platform = AndroidPlatform()
