+++ ohs-player-reference-app/src/androidMain/res/drawable-v24/ic_launcher_foreground.xml
