+++ ohs-player-reference-app/src/androidMain/res/mipmap-mdpi/ic_launcher.png
