+++ ohs-player-reference-app/src/androidMain/res/mipmap-xxhdpi/ic_launcher.png
