+++ ohs-player-reference-app/src/androidMain/res/mipmap-xxxhdpi/ic_launcher.png
