+++ ohs-player-reference-app/src/androidMain/res/values/strings.xml
