+++ ohs-player-reference-app/src/commonMain/composeResources/drawable/compose-multiplatform.xml
