+++ ohs-player-reference-app/src/commonMain/kotlin/dev/ohs/player/reference/app/Greeting.kt
@@ -1,0 +1,9 @@
+package dev.ohs.player.reference.app
+
+class Greeting {
+    private val platform = getPlatform()
+
+    fun greet(): String {
+        return "Hello, ${platform.name}!"
+    }
+}
