+++ ohs-player-reference-app/src/commonMain/kotlin/dev/ohs/player/reference/app/Platform.kt
@@ -1,0 +1,7 @@
+package dev.ohs.player.reference.app
+
+interface Platform {
+    val name: String
+}
+
+expect fun getPlatform(): Platform
