+++ ohs-player-reference-app/src/commonTest/kotlin/dev/ohs/player/reference/app/ComposeAppCommonTest.kt
@@ -1,0 +1,12 @@
+package dev.ohs.player.reference.app
+
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class ComposeAppCommonTest {
+
+    @Test
+    fun example() {
+        assertEquals(3, 1 + 2)
+    }
+}
