+++ ohs-player-reference-app/src/iosMain/kotlin/dev/ohs/player/reference/app/Platform.ios.kt
@@ -1,0 +1,9 @@
+package dev.ohs.player.reference.app
+
+import platform.UIKit.UIDevice
+
+class IOSPlatform: Platform {
+    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
+}
+
+actual fun getPlatform(): Platform = IOSPlatform()
