+++ ohs-player-reference-app/src/jsMain/kotlin/dev/ohs/player/reference/app/Platform.js.kt
@@ -1,0 +1,7 @@
+package dev.ohs.player.reference.app
+
+class JsPlatform: Platform {
+    override val name: String = "Web with Kotlin/JS"
+}
+
+actual fun getPlatform(): Platform = JsPlatform()
