+++ ohs-player-reference-app/src/jvmMain/kotlin/dev/ohs/player/reference/app/Platform.jvm.kt
@@ -1,0 +1,7 @@
+package dev.ohs.player.reference.app
+
+class JVMPlatform: Platform {
+    override val name: String = "Java ${System.getProperty("java.version")}"
+}
+
+actual fun getPlatform(): Platform = JVMPlatform()
