+++ ohs-player-reference-app/src/jvmMain/kotlin/dev/ohs/player/reference/app/main.kt
@@ -1,0 +1,13 @@
+package dev.ohs.player.reference.app
+
+import androidx.compose.ui.window.Window
+import androidx.compose.ui.window.application
+
+fun main() = application {
+    Window(
+        onCloseRequest = ::exitApplication,
+        title = "OHS Player Reference App",
+    ) {
+        App()
+    }
+}
