+++ ohs-player-reference-app/src/wasmJsMain/kotlin/dev/ohs/player/reference/app/Platform.wasmJs.kt
@@ -1,0 +1,7 @@
+package dev.ohs.player.reference.app
+
+class WasmPlatform: Platform {
+    override val name: String = "Web with Kotlin/Wasm"
+}
+
+actual fun getPlatform(): Platform = WasmPlatform()
