+++ ohs-player-reference-app/src/webMain/kotlin/dev/ohs/player/reference/app/main.kt
@@ -1,0 +1,11 @@
+package dev.ohs.player.reference.app
+
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.window.ComposeViewport
+
+@OptIn(ExperimentalComposeUiApi::class)
+fun main() {
+    ComposeViewport {
+        App()
+    }
+}
