+++ settings.gradle.kts
@@ -32,5 +32,5 @@ plugins {
     id("org.gradle.toolchains.foojay-resolver-convention") version "1.0.0"
 }
 
-include(":client-app")
+include(":ohs-player-reference-app")
 include(":ohs-player-library")
